refactor: update import paths for BaseGameProps and create shared types file #44
