+++ examples/mini-app/src/components/game/shared/types.ts
